requirements.txt
Django>=5.1.3
python-telegram-bot>=20.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
